lat lng float to double

+++ app/src/main/java/com/example/explorelens/data/model/PointOfIntrests/GeoLocation.kt
@@ -3,6 +3,6 @@ package com.example.explorelens.data.model.PointOfIntrests
 import com.google.gson.annotations.SerializedName
 
 data class GeoLocation (
-    @SerializedName("lat")  val lat: Float,
-    @SerializedName("lng") val lng: Float,
+    @SerializedName("lat")  val lat: Double,
+    @SerializedName("lng") val lng: Double,
 )
